步数转距离使用int64



See merge request !106

+++ util/math.go
@@ -10,7 +10,7 @@ func Round(f float64) int {
 }
 
 // StepsToDistance 步数转换为距离
-func StepsToDistance(steps int) float64 {
+func StepsToDistance(steps int64) float64 {
 	return float64((steps * 50 * (165 - 132) / 2700))
 	// calories = 70 * distance  * 1036 / 1000 / 1000;
 	// return float64(steps) * 0.6
